requirements.txt
Flask
kivy
requests
flask_sqlalchemy
mysql-connector-python
python-dotenv